Fix: cast downtime replica annotations
To avoid false conditions between integers and strings.

When setting arg downtime-replicas to 1, and setting annotation
downtime-replicas to 0 for a specific deployment, deployment was not
scaled-up to original replicas during uptime period.

This behaviour is caused when getting annotations as strings, and
comparing those values to integers, p.e.: `replicas ==
downtime_replicas` is false if types are different.